refactor(frontend): rename Ethereum components to add a new token (#4450)
# Motivation

Since we are going to create similar components for Solana network, it
makes sense to rename the Ethereum components that treat a new
additional token.
