+++ src/frontend/src/eth/components/tokens/EthAddTokenForm.svelte
