+++ src/frontend/src/eth/components/tokens/EthAddTokenReview.svelte
